refactor: remove default exports, remove utils, move Input
