+++ src/Button/index.js
@@ -54,4 +54,3 @@ Button.propTypes = {
 }
 
 export { Button }
-export default Button
