+++ src/Chip/index.js
@@ -128,4 +128,3 @@ Chip.propTypes = {
 }
 
 export { Chip }
-export default Chip
