+++ src/Logo/index.js
@@ -100,5 +100,3 @@ Logo.propTypes = {
 }
 
 export const LogoWhite = () => <Logo fill1="#ffffff" fill2="#ffffff" />
-
-export default Logo
