+++ src/SplitButton/index.js
@@ -106,4 +106,3 @@ SplitButton.propTypes = {
 }
 
 export { SplitButton }
-export default SplitButton
