--- src/utils/animations/index.js
@@ -1,3 +1,0 @@
-export * from './animatedScrollTo'
-export * from './easing'
-export * from './AnimatedCloseHOC'
