--- src/utils/css.js
@@ -1,5 +1,0 @@
-import classes from 'classnames/bind'
-
-export default function classNames(styles) {
-    return classes.bind(styles)
-}
