--- src/utils/html.js
@@ -1,7 +1,0 @@
-/**
- * get <html> or the <body> element in document
- * @returns {HTMLElement}
- */
-export function getDoc() {
-    return document.documentElement || document.body
-}
