--- src/utils/index.js
@@ -1,6 +1,0 @@
-export * from './html'
-export * from './url'
-export * from './css'
-export * from './rtl'
-export * from './timings'
-export * from './animations/index'
