--- src/utils/url.js
@@ -1,7 +1,0 @@
-/**
- * set window.location to path argument
- * @param path
- */
-export function gotoURL(path) {
-    window.location = path
-}
